+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Forward SDK info for sample profiles to Kafka. ([#507](https://github.com/getsentry/vroom/pull/507))
 - Forward SDK info for legacy profiles to Kafka. ([#515](https://github.com/getsentry/vroom/pull/515))
 - Add more metadata fields to Chunk Kafka message. ([#518](https://github.com/getsentry/vroom/pull/518))
+- Ingest Android profile chunks. ([#521](https://github.com/getsentry/vroom/pull/521))
 
 **Bug Fixes**: